+++ src/components/createQuestionnaire/utils.ts
@@ -18,8 +18,8 @@ import TableContainer from '@formcomponents/table/TableContainer';
 
 import { ComponentType } from 'react';
 import { QuestionnaireComponentItemProps } from '@/components/createQuestionnaire/GenerateQuestionnaireComponents';
-import { isTableCode } from '@/util';
-import { getItemWithIdFromResponseItemArray, getRootQuestionnaireResponseItemFromData, Path } from '@/util/refero-core';
+import { isRepeat, isTableCode } from '@/util';
+import { createIdSuffix, getItemWithIdFromResponseItemArray, getRootQuestionnaireResponseItemFromData, Path } from '@/util/refero-core';
 import { FormData } from '@/reducers/form';
 import ItemType from '@/constants/itemType';
 
@@ -129,4 +129,20 @@ export function getComponentForItem(
     return Quantity;
   }
   return undefined;
+}
+
+export const createIdFormComponentIds = (item?: QuestionnaireItem, path?: Path[], index?: number): string =>
+  encodeString(`${item?.linkId}${createIdSuffix(path, index, isRepeat(item))}`);
+
+export function encodeString(input: string): string {
+  // Replace all occurrences of "." with "%2E"
+  const encodedString = input.replace(/\./g, '%2E');
+
+  return encodedString;
+}
+
+export function decodeString(encodedId: string): string {
+  const originalString = encodedId.replace(/%2E/g, '.');
+
+  return originalString;
 }