+++ CHANGES.md
@@ -1,3 +1,12 @@
+## 16.0.0-beta.54
+- update default form values
+
+## 16.0.0-beta.53
+- Change the way id's are created. encode periods (.) to %2E to fix weird behaviour with react hook form
+
+## 16.0.0-beta.52
+- Bugfix: Activation of a repeatable group
+
 ## 16.0.0-beta.51
 - Bugfix: Fixed extra-label placement
 - Bugfix: Couldn't add zero to decimal field